requirements.txt
torch~=2.3.1
torchvision~=0.18.1
scikit-learn~=1.5.1
tqdm~=4.66.4
pyyaml~=6.0.1
pandas~=2.2.2
numpy~=1.26.4
pillow~=10.4.0
safetensors~=0.4.3
matplotlib~=3.9.1